requirements.txt
# Core ML
torch==2.2.0
torchvision==0.17.0
transformers==4.40.0
sentence-transformers==2.7.0

# Vision
Pillow==10.3.0
timm==0.9.16

# Data
datasets==2.19.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# Vector Search
faiss-cpu==1.8.0

# MLOps
mlflow==2.12.1
wandb==0.17.0

# Experiment config
pyyaml==6.0.1
omegaconf==2.3.0

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
umap-learn==0.5.6

# App
streamlit==1.34.0

# Utilities
tqdm==4.66.4
requests==2.31.0
huggingface_hub==0.23.0